Task-Manager
The manager is assigned to the task personnel by the manager according to the Priority and Urgency information (Task's title, content, priority and urgency conditions, completion time, if any, by adding the documents or documents related to the task, by selecting the relevant personnel) and in the background by e-mail. This information goes to the staff.

Assigned Tasks are assigned to the personnel's screen, taking into account the priority and urgency specified.
My project that I developed using Asp.Net and I use MsSQL in the database.